Split into two listing components
