+++ resources/js/addon.js
@@ -1,7 +1,9 @@
 import MuxMirrorFieldtype from './components/MuxMirrorFieldtype.vue'
 import MuxMirrorIndexFieldtype from './components/MuxMirrorIndexFieldtype.vue'
-import MuxVideoListing from './components/MuxVideoListing.vue'
+import MuxMirroredListing from './components/MuxMirroredListing.vue'
+import MuxLibraryListing from './components/MuxLibraryListing.vue'
 
 Statamic.$components.register('mux_mirror-fieldtype', MuxMirrorFieldtype)
 Statamic.$components.register('mux_mirror-fieldtype-index', MuxMirrorIndexFieldtype)
-Statamic.$components.register('mux-video-listing', MuxVideoListing)
+Statamic.$components.register('mux-mirrored-listing', MuxMirroredListing)
+Statamic.$components.register('mux-library-listing', MuxLibraryListing)
