+++ resources/views/cp/library.blade.php
@@ -1,0 +1,9 @@
+@extends('statamic::layout')
+
+@section('content')
+    <mux-library-listing
+        remote-endpoint="{{ $endpoint }}"
+        refresh-endpoint="{{ $refreshEndpoint }}"
+        dashboard-url="{{ $dashboardUrl }}"
+    />
+@endsection
