+++ resources/views/cp/mirrored.blade.php
@@ -1,0 +1,8 @@
+@extends('statamic::layout')
+
+@section('content')
+    <mux-mirrored-listing
+        local-endpoint="{{ $endpoint }}"
+        command-endpoint="{{ $commandEndpoint }}"
+    />
+@endsection
