+++ src/ServiceProvider.php
@@ -171,6 +171,7 @@ class ServiceProvider extends AddonServiceProvider
     {
         Nav::extend(function (\Statamic\CP\Navigation\Nav $nav) {
             $nav->tools('Mux')
+                ->route('mux.index')
                 ->icon('fieldtype-video')
                 ->can('view mux')
                 ->children([
